requirements.txt
Flask==3.0.3
Markdown==3.6
python-frontmatter==1.0.1
PyYAML==6.0.2
bleach==6.1.0
